requirements.txt
pylint
pytest
pytest-cov
click
python-dotenv
langchain
chromadb==0.3.10
tiktoken
openai
black
metapub
pandas
streamlit
streamlit-chat==0.0.2.2